+++ xma/layers_jax/depthwise_causal_convolution/jax_implementation.py
@@ -1,0 +1,47 @@
+# **************************************************
+# Copyright (c) 2026, Mayank Mishra
+# **************************************************
+
+import jax
+import jax.numpy as jnp
+
+
+def _depthwise_causal_convolution_reference(
+    x: jax.Array, W: jax.Array, b: jax.Array | None, h0: jax.Array | None, output_state: bool
+) -> tuple[jax.Array, jax.Array | None]:
+    _, S, H = x.shape
+    K = W.shape[-1]
+
+    x = jnp.transpose(x, (0, 2, 1))
+    ht = None
+
+    if h0 is None:
+        padding = [(K - 1, 0)]
+
+        if output_state:
+            ht = jnp.pad(x, ((0, 0), (0, 0), (K - S, 0))) if S < K else x[:, :, -K:]
+    else:
+        padding = [(0, 0)]
+
+        x = jnp.concatenate([h0.astype(x.dtype), x], axis=-1)
+
+        if output_state:
+            ht = x[:, :, -K:]
+
+        x = x[:, :, 1:]
+
+    x = jax.lax.conv_general_dilated(
+        lhs=x,
+        rhs=W[:, None, :],
+        window_strides=(1,),
+        padding=padding,
+        feature_group_count=H,
+        dimension_numbers=("NCH", "OIH", "NCH"),
+    )
+
+    if b is not None:
+        x = x + b[None, :, None]
+
+    x = jnp.transpose(x, (0, 2, 1))
+
+    return x, ht
